backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pyjwt
passlib[bcrypt]
pydantic-settings